refactor: 페이지 내 위젯 정의 - header footer manager
