+++ src/pages/PostManagerPage.tsx
@@ -1,0 +1,11 @@
+import { PostManagerWidget } from '../widgets/post-manager/PostManagerWidget';
+
+const PostManagerPage = () => {
+  return (
+    <div>
+      <PostManagerWidget />
+    </div>
+  );
+};
+
+export default PostManagerPage;
