+++ src/widgets/footer/ui/Footer.tsx
